requirements.txt
opencv-python
heapq
collections 
deque
numpy 
matplotlib.pyplot 
random
cv2
shapely